fix: replace int with size_t in loops

+++ 03/K-Floyd.cpp
@@ -13,8 +13,8 @@ int main() {
 
     vector<vector<int>> matrix(n, vector<int>(n));
 
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < n; ++j) {
+    for (size_t i = 0; i < n; ++i) {
+        for (size_t j = 0; j < n; ++j) {
             int tmp;
             cin >> tmp;
             matrix[i][j] = tmp;
@@ -22,22 +22,22 @@ int main() {
     }
 
     vector<vector<int>> res(n, vector<int>(n));
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < n; ++j) {
+    for (size_t i = 0; i < n; ++i) {
+        for (size_t j = 0; j < n; ++j) {
             res[i][j] = matrix[i][j];
         }
     }
-    
-    for (int i = 0; i < n; ++i) {
-        for (int u = 0; u < n; ++u) {
-            for (int v = 0; v < n; ++v) {
+
+    for (size_t i = 0; i < n; ++i) {
+        for (size_t u = 0; u < n; ++u) {
+            for (size_t v = 0; v < n; ++v) {
                 res[u][v] = min(res[u][v], res[u][i] + res[i][v]);
             }
         }
     }
 
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < n; ++j) {
+    for (size_t i = 0; i < n; ++i) {
+        for (size_t j = 0; j < n; ++j) {
             cout << res[i][j] << " ";
         }
         cout << endl;
